+++ mono_repo/test/presubmit_command_test.dart
@@ -25,7 +25,8 @@ void main() {
           () => presubmit(rootDirectory: d.sandbox),
           throwsUserExceptionWith(
               'No $travisShPath file found, please run the `travis` '
-              'command first.'));
+              'command first.',
+              isNull));
     });
   });